requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
pydantic[email]
pydantic-settings
httpx
pandas
google-generativeai
jinja2
aiofiles
reportlab
openpyxl
beautifulsoup4
fastapi-mail
